chore: move fork errors into relevant files for clarity, split out abi decode into seperate file for clarity
