+++ crates/common/src/lib.rs
@@ -1,8 +1,9 @@
 pub mod add_order;
+pub mod decode_abi_error;
 pub mod deposit;
-pub mod error;
 pub mod fork;
 pub mod remove_order;
 pub mod subgraph;
 pub mod transaction;
+pub mod transaction_error;
 pub mod withdraw;
